Refactor: Moved pages to new folder + routing fix
