--- src/client/plugin/components/IgnoredPage.tsx
@@ -1,8 +1,0 @@
-export default function IgnoredPage() {
-  return (
-    <div className="rounded-lg shadow p-6">
-      <h2 className="text-xl font-semibold mb-4">Ignored Conflicts</h2>
-      <p className="">Manage your ignored conflicts here.</p>
-    </div>
-  );
-}
