--- src/client/plugin/components/MainPage.tsx
@@ -1,8 +1,0 @@
-export default function MainPage() {
-  return (
-    <div className="rounded-lg shadow p-6">
-      <h2 className="text-xl font-semibold mb-4">Main Dashboard</h2>
-      <p className="text-black">Welcome to the Schedule Builder!</p>
-    </div>
-  );
-}
