--- src/client/plugin/components/SettingsPage.tsx
@@ -1,8 +1,0 @@
-export default function SettingsPage() {
-  return (
-    <div className="rounded-lg shadow p-6">
-      <h2 className="text-xl font-semibold mb-4">Settings</h2>
-      <p className="">Configure your application settings here.</p>
-    </div>
-  );
-}
