+++ src/client/plugin/contexts/themeContext.tsx
@@ -5,8 +5,6 @@ interface ThemeContextType {
   currentTheme: ColorTheme;
   setTheme: (theme: ColorTheme) => void;
   predefinedThemes: ColorTheme[];
-  isSettingsOpen: boolean;
-  setIsSettingsOpen: (open: boolean) => void;
 }
 
 const themeContext = createContext<ThemeContextType | undefined>(undefined);
